+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.1.12
 * Added new Metadata API - List categories (v2)
+* Added new Team API - Get Work Diary by Contract)
 
 ## 0.1.11
 * Added new Milestone API - Get Active Milestone for specific Contract